! fixed a case where would fail to deploy a never before deployed windmill.
Note: Made sure `NoResultException` is not rolled back and is considered ApplicationException(s). Previously this would cause the transaction to be marked as rolled back ("java.sql.SQLException: IJ031070: Transaction cannot proceed: STATUS_MARKED_ROLLBACK") when trying to deploy a new app. The code is designed to recover from a `NoResultException` exception. In fact, it's part of the "deploy a new windmill" contract.

~ updated jdbc driver to the latest version